buncha style adjustments
