--- frontend/components/engine-badge/engine-badge.module.css
@@ -1,23 +1,0 @@
-.wrapper {
-	overflow: hidden;
-	border-radius: var(--mantine-radius-default);
-	text-align: center;
-}
-
-.brand {
-	font-weight: bold;
-	text-transform: uppercase;
-}
-
-.version {
-	background-color: rgba(0, 0, 0, 0.3);
-	font-family: monospace;
-	font-size: var(--mantine-font-size-xs);
-	text-transform: lowercase;
-}
-
-.details {
-	font-size: 10px;
-	opacity: 0.5;
-	text-transform: lowercase;
-}
